Break out the BagReader tests into test cases
