+++ bag_replicator/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator/services/BagReplicatorWorkerTest.scala
@@ -15,7 +15,7 @@ import uk.ac.wellcome.messaging.worker.models.{
 import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.BagReplicatorFixtures
 import uk.ac.wellcome.platform.archive.bagreplicator.models.ReplicationSummary
 import uk.ac.wellcome.platform.archive.common.EnrichedBagInformationPayload
-import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
+import uk.ac.wellcome.platform.archive.common.fixtures.S3BagLocationFixtures
 import uk.ac.wellcome.platform.archive.common.generators.PayloadGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.fixtures.IngestUpdateAssertions
 import uk.ac.wellcome.storage.locking.{LockDao, LockFailure}
@@ -28,7 +28,7 @@ import scala.concurrent.Future
 class BagReplicatorWorkerTest
     extends FunSpec
     with Matchers
-    with BagLocationFixtures
+    with S3BagLocationFixtures
     with BagReplicatorFixtures
     with IngestUpdateAssertions
     with PayloadGenerators
