+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/bagit/services/s3/S3BagReaderTest.scala
@@ -1,0 +1,30 @@
+package uk.ac.wellcome.platform.archive.common.bagit.services.s3
+
+import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.platform.archive.common.bagit.services.{BagReader, BagReaderTestCases}
+import uk.ac.wellcome.storage.ObjectLocation
+import uk.ac.wellcome.storage.fixtures.S3Fixtures
+import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
+import uk.ac.wellcome.storage.store.TypedStore
+import uk.ac.wellcome.storage.store.s3.{S3StreamStore, S3TypedStore}
+
+class S3BagReaderTest extends BagReaderTestCases[Bucket] with S3Fixtures {
+  override def withTypedStore[R](testWith: TestWith[TypedStore[ObjectLocation, String], R]): R = {
+    implicit val s3StreamStore: S3StreamStore = new S3StreamStore()
+
+    testWith(new S3TypedStore[String]())
+  }
+
+  override def withNamespace[R](testWith: TestWith[Bucket, R]): R =
+    withLocalS3Bucket { bucket =>
+      testWith(bucket)
+    }
+
+  override val bagReader: BagReader[_] = new S3BagReader()
+
+  override def deleteFile(rootLocation: ObjectLocation, path: String): Unit =
+    s3Client.deleteObject(
+      rootLocation.namespace,
+      rootLocation.join(path).path
+    )
+}
